Refactored for better file structure.
