+++ static/styles.css
@@ -7,7 +7,7 @@ h2 {
 }
 
 h3 {
-    color: black;
+    color: rgb(231, 14, 14);
 }
 
 body {
