+++ templates/courses/course_history.html
