+++ templates/courses/joint_courses.html
