+++ templates/faculty/faculty_list.html
@@ -1,0 +1,40 @@
+<html>
+<head>
+    <title>Faculty Code Finder</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
+    <!-- <link rel="stylesheet" href="/static/styles.css"> -->
+</head>
+<body>
+    <div class="container-fluid">
+        <div>
+            <p><a href="/">Menu</a></p>
+        </div>
+        <div>
+            <h3>Instructor List</h3>
+            <!-- <ul>
+            {% for eid, instructor, netid in instructor_list%}
+                <li>{{ instructor }} - {{ eid }} - {{ netid }}</li>
+            {% endfor %}
+            </ul> -->
+            <table class="table w-auto table table-sm table-bordered table-striped">
+                <thead class="thead-light">
+                    <tr>
+                        <th>Name</th>
+                        <th>EID</th>
+                        <th>NetID</th>
+                    </tr>
+                </thead>
+                <tbody>
+                {% for eid, instructor, netid in instructor_list %}
+                    <tr>
+                        <td> {{ instructor }} </td>
+                        <td> {{ eid }} </td>
+                        <td> {{ netid }} </td>
+                    </tr>
+                {% endfor %}
+                </tbody>
+            </table>
+        </div>
+    </div>
+</body>
+</html>
