--- templates/faculty_list.html
@@ -1,21 +1,0 @@
-<html>
-<head>
-    <title>Faculty Code Finder</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
-</head>
-<body>
-    <div class="container-fluid">
-        <div>
-            <p><a href="/">Menu</a></p>
-        </div>
-        <div>
-            <h3>Instructor List</h3>
-            <ul>
-            {% for eid, instructor, netid in instructor_list%}
-                <li>{{ instructor }} - {{ eid }} - {{ netid }}</li>
-            {% endfor %}
-            </ul>
-        </div>
-    </div>
-</body>
-</html>
