--- templates/item.html
@@ -1,9 +1,0 @@
-<html>
-<head>
-    <title>Item Details</title>
-    <link href={{ url_for("static", path='/styles.css') }} rel='stylesheet'>
-</head>
-<body>
-    <h1>Item ID: {{ id }}</h1>
-</body>
-</html>
