+++ templates/students/student_info.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 {% block title %}Student Info{% endblock %}
 {% block body %}
     <div class="container-fluid">
